+++ api-server/source/notifications/notification-routes.ts
@@ -1,4 +1,5 @@
 import { getWebPushCredentials } from "./get-web-push-credentials.ts";
+import { testWebPushRoute } from "./test-web-push.ts";
 import { createWebPushDevicesRoute } from "./web-push-device-create.ts";
 import { deleteWebPushDevicesRoute } from "./web-push-device-delete.ts";
 import { updateWebPushDevicesRoute } from "./web-push-device-update.ts";
@@ -10,4 +11,5 @@ export const notificationRoutes = [
   deleteWebPushDevicesRoute,
   listWebPushDevicesRoute,
   updateWebPushDevicesRoute,
+  testWebPushRoute,
 ];