+++ api-server/CHANGELOG.md
@@ -2,10 +2,13 @@
 
 This file documents changes to the Deconf API Server.
 
-## next
+## 0.3.1
+
+- **new** web-push admin endpoints to test & broadcast
+
+## 0.3.0
 
 - **new** web-push "test" endpoint
-- **new** web-push admin endpoints to test & broadcast
 - **new** notify command to run notifications process
 - **new** sign in with Google
 - **fix** align web-push endpoints